requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.16.2
python-dotenv==1.0.1
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
langchain-groq==0.2.0
langchain-core==0.3.12
apscheduler==3.10.4
jinja2==3.1.5
streamlit==1.39.0
google-api-python-client==2.153.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
pandas==2.2.3
python-dateutil==2.9.0
pydantic-settings
